[addons] Make CRepositoryUpdateJob internal for CRepositoryUpdater.
